New executive compensation response (#1166)
* Expect executive comp fields to be moved from recipient.officers to latest_transaction.contract_data

* Updated tests and mock responses
