+++ src/js/components/award/details/additional/ContractAdditionalDetails.jsx
@@ -50,7 +50,7 @@ export default class ContractAdditionalDetails extends React.Component {
                     title="Additional Details" />
                 <hr className="additional-group-divider" />
                 <AdditionalGroup
-                    data={award.recipient.officers}
+                    data={award.additionalDetails.officers}
                     fields={DataFields.compensationFields}
                     title="Executive Compensation" />
             </div>
