+++ src/js/models/v2/awards/BaseAwardRecipient.js
@@ -4,7 +4,6 @@
  */
 
 import { getBusinessTypes } from 'helpers/businessTypesHelper';
-import { formatMoney } from 'helpers/moneyFormatter';
 import CoreLocation from 'models/v2/CoreLocation';
 
 const parseBusinessCategories = (data) => (
@@ -15,28 +14,6 @@ const parseBusinessCategories = (data) => (
         return parsed;
     }, [])
 );
-
-const parseExecutiveCompensation = (data) => {
-    const executiveCompensation = {};
-    if (data) {
-        for (let i = 1; i < 6; i++) {
-            const name = data[`officer_${i}_name`] || '';
-            const amount = formatMoney(data[`officer_${i}_amount`]) || 0;
-            if (name) {
-                executiveCompensation[`officer${i}`] = `${name} - ${amount}`;
-            }
-            else {
-                executiveCompensation[`officer${i}`] = '--';
-            }
-        }
-    }
-    else {
-        for (let i = 1; i < 6; i++) {
-            executiveCompensation[`officer${i}`] = '--';
-        }
-    }
-    return executiveCompensation;
-};
 
 const BaseAwardRecipient = {
     populate(data) {
@@ -69,15 +46,6 @@ const BaseAwardRecipient = {
         const location = Object.create(CoreLocation);
         location.populateCore(locationData);
         this.location = location;
-
-        // Executive Compensation
-        if (data.officers) {
-            this.officers = parseExecutiveCompensation(data.officers);
-        }
-        else {
-            // data.officers is undefined
-            this.officers = parseExecutiveCompensation(null);
-        }
     },
     get businessTypes() {
         if (this._businessTypes) {
