+++ tests/models/awards/additionalDetails/BaseContractAdditionalDetails-test.js
@@ -3,9 +3,8 @@
  * Created by Lizzie Salita 3/20/18
  */
 
-
-import { mockContractApi } from '../mockAwardApi';
 import BaseContractAdditionalDetails from 'models/v2/awards/additionalDetails/BaseContractAdditionalDetails';
+import { mockContractApi } from '../mockAwardApi';
 
 const details = Object.create(BaseContractAdditionalDetails);
 details.populate(mockContractApi.latest_transaction.contract_data);
@@ -19,5 +18,14 @@ describe('BaseContractAdditionalDetails', () => {
     });
     it('should return -- for null values', () => {
         expect(details.clingerCohenAct).toEqual('--');
+    });
+    it('should parse executive compensation', () => {
+        expect(details.officers).toEqual({
+            officer1: 'George Washington - $9,000',
+            officer2: 'John Adams - $7,001',
+            officer3: 'Thomas Jefferson - $6,000',
+            officer4: 'James Madison - $5,000',
+            officer5: '--'
+        });
     });
 });
